fix(server): extract after:mutation tests to tests.rs (#460)
The `routes::after_mutation` module shipped its unit tests as an inline
`#[cfg(test)] mod tests { … }` block, which trips the `lint-tests-layout`
preflight gate ("inline test blocks found — extract to tests.rs"). Split the
single file into `after_mutation/mod.rs` (the planner + gated dispatch) and
`after_mutation/tests.rs`, matching the workspace convention. No behavior change.

## Verification
✅ make lint-tests-layout (+ unwrap/expect/async-trait/gate-db/gate-core/
   check-test-imports/check-route-syntax/check-deploy-security/
   check-audit-lockstep/check-deadlines/test-release-tooling)
✅ cargo +nightly fmt --check
✅ cargo test -p fraiseql-server --lib after_mutation (10 passed)
